api/requirements.txt
fastapi==0.110.1
httpx==0.27.0
uvicorn==0.29.0
pyyaml==6.0.1
